+++ javascript/src/local/index.ts
@@ -1,2 +1,3 @@
 export * from "./core.js";
 export * from "./context.js";
+export * from "./tools.js";